+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fix: include/exclude matches against full path
 - Fix: include/exclude pattern matching differs from rsync
 
 ## 1.7.0 - 2021-11-24